backend/requirements.txt
amqp==5.3.1
asgiref==3.9.2
barcode==1.0.4
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
Levenshtein==0.27.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyPDF2==3.0.1
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-Levenshtein==0.27.1
qrcode==8.2
RapidFuzz==3.14.2
razorpay==2.0.0
redis==7.0.1
reportlab==4.4.4
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
